First code dump for the new sgp4 polyjectory support.
Testing, cleanups, docs, validation, etc. all to be done yet.

+++ mizuba/data_sources/_common.py
@@ -21,6 +21,7 @@ from typing import Any, Tuple
 
 def _common_validate_gpes(gpes: pl.DataFrame, unique_norad_id: bool = True) -> None:
     # Common logic for the validation of GPEs downloaded from space-track.org or celestrak.org.
+    import polars as pl
 
     # We need all GPEs to have a non-null norad id.
     if not gpes["NORAD_CAT_ID"].is_not_null().all():
@@ -28,7 +29,7 @@ def _common_validate_gpes(gpes: pl.DataFrame, unique_norad_id: bool = True) -> N
 
     # If unique_norad_id is True, check that the norad ids
     # are unique.
-    if unique_norad_id and not gpes["NORAD_CAT_ID"].cast(int).is_unique().all():
+    if unique_norad_id and not gpes["NORAD_CAT_ID"].cast(pl.UInt64).is_unique().all():
         raise ValueError("Non-unique NORAD IDs detected in GPEs")
 
     # Check that all the data used for orbital propagation is present.
